chore: move test rt to mod test
