+++ benches/impl_path_string_for_evaluation_context.rs
@@ -20,7 +20,7 @@ use tailcall::cache::InMemoryCache;
 use tailcall::http::{RequestContext, Response};
 use tailcall::lambda::{EvaluationContext, ResolverContextLike};
 use tailcall::path::PathString;
-use tailcall::runtime::TargetRuntime;
+use tailcall::target_runtime::TargetRuntime;
 use tailcall::{EnvIO, FileIO, HttpIO};
 
 struct Http {
