--- src/runtime/mod.rs
@@ -1,2 +1,0 @@
-pub mod target_runtime;
-pub use target_runtime::*;
